+++ Bridgeport/__init__.py
@@ -1,0 +1,3 @@
+from chimpss.bridgeport import Bridgeport
+
+__all__ = ["Bridgeport"]