+++ src/chimpss/bridgeport/__init__.py
@@ -1,0 +1,3 @@
+from chimpss.bridgeport.bridgeport import Bridgeport
+
+__all__ = ["Bridgeport"]